Add a way to visualize a line running along bones.
This is eventually intended to be generated from the bone rotation,
to obtain information we discard when rendering as a cylinder.

For now, it's constantly rotating, for demonstration purposes.
There's some interesting artifacting going on here, I'd like to stick
the skeleton in a single spot so that I can try to more closely observe
what's going on.