Switch to standing universe, which is "close enough"
Driver and Feeder now use a "static standing" universe, which should be
the same as standing universe except it ignores playspace changing.

Ideally, we'd implement that here too, but this is better than leaving
it the way it is. When playspace movement happens, I expect the skeleton
to move relative to the player.